+++ troposphere/ecr.py
@@ -6,10 +6,18 @@ except ImportError:
     policytypes = dict,
 
 
+class LifecyclePolicy(AWSObject):
+    props = {
+        'LifecyclePolicyText': (basestring, False),
+        'RegistryId': (basestring, False),
+    }
+
+
 class Repository(AWSObject):
     resource_type = "AWS::ECR::Repository"
 
     props = {
+        'LifecyclePolicy': (basestring, False),
         'RepositoryName': (basestring, False),
         'RepositoryPolicyText': (policytypes, False),
     }