+++ troposphere/serverless.py
@@ -5,7 +5,7 @@
 
 import types
 
-from . import AWSObject, AWSProperty, Tags
+from . import AWSObject, AWSProperty
 from .awslambda import Environment, VPCConfig, validate_memory_size
 from .dynamodb import ProvisionedThroughput
 from .validators import exactly_one, positive_integer
@@ -68,7 +68,7 @@ class Function(AWSObject):
         'Environment': (Environment, False),
         'VpcConfig': (VPCConfig, False),
         'Events': (dict, False),
-        'Tags': (Tags, False),
+        'Tags': (dict, False),
         'Tracing': (basestring, False),
         'KmsKeyArn': (basestring, False),
         'DeadLetterQueue': (DeadLetterQueue, False)