+++ troposphere/ssm.py
@@ -78,7 +78,7 @@ class Rule(AWSProperty):
     props = {
         'ApproveAfterDays': (integer, False),
         'ComplianceLevel': (compliance_level, False),
-        'PatchFilterGroup': (PatchFilter, False),
+        'PatchFilterGroup': (PatchFilterGroup, False),
     }